requirements.txt
dashscope
requests
pyyaml
